Make CUDA isolated-site and isolated-GBSA forces precision-generic
IsolatedSite computed the center of mass in single precision regardless of context precision; accumulate it in real and pass the site geometry as double so the restraint follows the selected precision.

IsolatedGBSA stored receptor positions and parameters as float and ran the descreening math in float, so it failed to compile in double precision and otherwise carried only single-precision accuracy. Store receptor data as real and run the HCT/Born/GB math in real. Double precision now agrees with the serial reference to ~1e-8; single and mixed are unchanged.

+++ platforms/cuda/src/CudaIsolatedGBSAKernels.h
@@ -49,7 +49,7 @@ private:
     // GB parameters
     IsolatedGBSAForce::GBMethod gbMethod;
     IsolatedGBSAForce::ReceptorMode receptorMode;
-    float prefactor;  // -138.935456 * (1/solute - 1/solvent)
+    double prefactor;  // -138.935456 * (1/solute - 1/solvent)
     bool includeSurfaceArea;
     float surfaceTension;
     float cutoffDistance;  // -1.0 for no cutoff
